+++ pkg/resource/deploy/step_generator.go
@@ -74,6 +74,7 @@ func (sg *stepGenerator) GenerateReadSteps(event ReadResourceEvent) ([]Step, res
 		false, /* deleteBeforeCreate */
 		event.AdditionalSecretOutputs(),
 		nil, /* aliases */
+		nil, /* customTimeouts */
 	)
 	old, hasOld := sg.plan.Olds()[urn]
 
@@ -163,7 +164,7 @@ func (sg *stepGenerator) GenerateSteps(event RegisterResourceEvent) ([]Step, res
 	// get serialized into the checkpoint file.
 	new := resource.NewState(goal.Type, urn, goal.Custom, false, "", inputs, nil, goal.Parent, goal.Protect, false,
 		goal.Dependencies, goal.InitErrors, goal.Provider, goal.PropertyDependencies, false,
-		goal.AdditionalSecretOutputs, goal.Aliases)
+		goal.AdditionalSecretOutputs, goal.Aliases, &goal.CustomTimeouts)
 
 	// Is this thing a provider resource? If so, stash it - we might need it later when calculating replacement
 	// of resources that use this provider.