+++ pkg/resource/resource_goal.go
@@ -36,15 +36,16 @@ type Goal struct {
 	AdditionalSecretOutputs []PropertyKey         // outputs that should always be treated as secrets.
 	Aliases                 []URN                 // additional URNs that should be aliased to this resource.
 	ID                      ID                    // the expected ID of the resource, if any.
+	CustomTimeouts          CustomTimeouts        // an optional config object for resource options
 }
 
 // NewGoal allocates a new resource goal state.
 func NewGoal(t tokens.Type, name tokens.QName, custom bool, props PropertyMap,
 	parent URN, protect bool, dependencies []URN, provider string, initErrors []string,
 	propertyDependencies map[PropertyKey][]URN, deleteBeforeReplace bool, ignoreChanges []string,
-	additionalSecretOutputs []PropertyKey, aliases []URN, id ID) *Goal {
+	additionalSecretOutputs []PropertyKey, aliases []URN, id ID, customTimeouts *CustomTimeouts) *Goal {
 
-	return &Goal{
+	g := &Goal{
 		Type:                    t,
 		Name:                    name,
 		Custom:                  custom,
@@ -61,4 +62,10 @@ func NewGoal(t tokens.Type, name tokens.QName, custom bool, props PropertyMap,
 		Aliases:                 aliases,
 		ID:                      id,
 	}
+
+	if customTimeouts != nil {
+		g.CustomTimeouts = *customTimeouts
+	}
+
+	return g
 }