+++ sdk/go/pulumi/resource.go
@@ -67,10 +67,18 @@ type ResourceOpt struct {
 	// current state. Once a resource has been imported, the import property must be removed from the resource's
 	// options.
 	Import ID
+	// CustomTimeouts is an optional configuration block used for CRUD operations
+	CustomTimeouts *CustomTimeouts
 }
 
 // InvokeOpt contains optional settings that control an invoke's behavior.
 type InvokeOpt struct {
 	// Provider is an optional provider resource to use for this invoke.
 	Provider ProviderResource
+}
+
+type CustomTimeouts struct {
+	Create string
+	Update string
+	Delete string
 }