+++ CHANGELOG.md
@@ -4,18 +4,25 @@ CHANGELOG
 ## HEAD (Unreleased)
 - `pulumi new` allows specifying a local path to templates (resolves
    [#2672](https://github.com/pulumi/pulumi/issues/2672))
+   
 - Fix an issue where a file archive created on Windows would contain back-slashes
   [#2784](https://github.com/pulumi/pulumi/issues/2784))
+
 - Fix an issue where output values of a resource would not be present when they
   contained secret values, when using Python.
+
 - Fix an issue where emojis are printed in non-interactive mode. (fixes
   [#2871](https://github.com/pulumi/pulumi/issues/2871))
 - Promises/Outputs can now be directly exported as the top-level (i.e. not-named) output of a Stack.
   (fixes [#2910](https://github.com/pulumi/pulumi/issues/2910))
+  
 - Add support for importing existing resources to be managed using Pulumi. A resource can be imported
   by setting the `import` property in the resource options bag when instantiating a resource. In order to
   successfully import a resource, its desired configuration (i.e. its inputs) must not differ from its
   actual configuration (i.e. its state) as calculated by the resource's provider.
+  
+- Add the ability to pass a customTimeouts object from the providers across the engine
+  to resource management. (fixes [#2665](https://github.com/pulumi/pulumi/issues/2665))
 
 ## 0.17.22 (2019-07-11)